+++ cli/codegen/generate_code.go
@@ -24,14 +24,6 @@ var luaCodeFiles = []generateLuaCodeOpts{
 		},
 	},
 	{
-		PackageName: "connect",
-		FileName:    "cli/connect/lua_code_gen.go",
-		VariablesMap: map[string]string{
-			"evalFuncBody":           "cli/connect/lua/eval_func_body.lua",
-			"getSuggestionsFuncBody": "cli/connect/lua/get_suggestions_func_body.lua",
-		},
-	},
-	{
 		PackageName: "connector",
 		FileName:    "cli/connector/lua_code_gen.go",
 		VariablesMap: map[string]string{